Unify query model in trace, now data source api like timeseries/list is supported
