+++ server/storage-jdbc/src/main/java/org/bithon/server/storage/jdbc/tracing/reader/TraceJdbcReader.java
@@ -34,7 +34,9 @@ import org.bithon.component.commons.utils.StringUtils;
 import org.bithon.server.commons.time.TimeSpan;
 import org.bithon.server.storage.datasource.ISchema;
 import org.bithon.server.storage.datasource.query.IDataSourceReader;
+import org.bithon.server.storage.datasource.query.Limit;
 import org.bithon.server.storage.datasource.query.Order;
+import org.bithon.server.storage.datasource.query.OrderBy;
 import org.bithon.server.storage.datasource.query.Query;
 import org.bithon.server.storage.jdbc.common.dialect.Expression2Sql;
 import org.bithon.server.storage.jdbc.common.dialect.ISqlDialect;
@@ -44,6 +46,7 @@ import org.bithon.server.storage.tracing.ITraceReader;
 import org.bithon.server.storage.tracing.TraceSpan;
 import org.bithon.server.storage.tracing.TraceStorageConfig;
 import org.bithon.server.storage.tracing.mapping.TraceIdMapping;
+import org.bithon.server.storage.tracing.reader.TraceFilterSplitter;
 import org.jooq.Condition;
 import org.jooq.DSLContext;
 import org.jooq.Field;
@@ -110,10 +113,8 @@ public class TraceJdbcReader implements ITraceReader {
                                         List<IExpression> indexedTagFilter,
                                         Timestamp start,
                                         Timestamp end,
-                                        String orderBy,
-                                        Order order,
-                                        int pageNumber,
-                                        int pageSize) {
+                                        OrderBy orderBy,
+                                        Limit limit) {
         boolean isOnSummaryTable = isFilterOnRootSpanOnly(filter);
 
         Field<LocalDateTime> timestampField = isOnSummaryTable ? Tables.BITHON_TRACE_SPAN_SUMMARY.TIMESTAMP : Tables.BITHON_TRACE_SPAN.TIMESTAMP;
@@ -146,22 +147,22 @@ public class TraceJdbcReader implements ITraceReader {
         }
 
         Field<?> orderField;
-        if ("costTime".equals(orderBy)) {
+        if ("costTime".equals(orderBy.getName())) {
             orderField = isOnSummaryTable ? Tables.BITHON_TRACE_SPAN_SUMMARY.COSTTIMEMS : Tables.BITHON_TRACE_SPAN.COSTTIMEMS;
-        } else if ("startTime".equals(orderBy)) {
+        } else if ("startTime".equals(orderBy.getName())) {
             orderField = isOnSummaryTable ? Tables.BITHON_TRACE_SPAN_SUMMARY.STARTTIMEUS : Tables.BITHON_TRACE_SPAN.COSTTIMEMS;
         } else {
             orderField = isOnSummaryTable ? Tables.BITHON_TRACE_SPAN_SUMMARY.TIMESTAMP : Tables.BITHON_TRACE_SPAN.COSTTIMEMS;
         }
         SelectSeekStep1<?, ?> orderedListQuery;
-        if (Order.desc.equals(order)) {
+        if (Order.desc.equals(orderBy.getOrder())) {
             orderedListQuery = listQuery.orderBy(orderField.desc());
         } else {
             orderedListQuery = listQuery.orderBy(orderField.asc());
         }
 
-        return dslContext.fetch(getSQL(orderedListQuery.offset(pageNumber * pageSize)
-                                                       .limit(pageSize)))
+        return dslContext.fetch(getSQL(orderedListQuery.offset(limit.getOffset())
+                                                       .limit(limit.getLimit())))
                          .map(this::toTraceSpan);
     }
 
@@ -170,7 +171,7 @@ public class TraceJdbcReader implements ITraceReader {
                                                           List<IExpression> indexedTagFilter,
                                                           Timestamp start,
                                                           Timestamp end,
-                                                          int interval) {
+                                                          long interval) {
         boolean isOnSummaryTable = isFilterOnRootSpanOnly(filter);
 
         String timeBucket = sqlDialect.timeFloorExpression(new IdentifierExpression("timestamp"), interval);
@@ -354,7 +355,14 @@ public class TraceJdbcReader implements ITraceReader {
 
     @Override
     public List<Map<String, Object>> timeseries(Query query) {
-        return getDataSourceReader().timeseries(query);
+        TraceFilterSplitter splitter = new TraceFilterSplitter(this.traceSpanSchema, this.traceTagIndexSchema);
+        splitter.split(query.getFilter());
+
+        return getTraceDistribution(splitter.getExpression(),
+                                    splitter.getIndexedTagFilters(),
+                                    query.getInterval().getStartTime().toTimestamp(),
+                                    query.getInterval().getEndTime().toTimestamp(),
+                                    query.getInterval().getStep().getSeconds());
     }
 
     @Override
@@ -363,13 +371,27 @@ public class TraceJdbcReader implements ITraceReader {
     }
 
     @Override
-    public List<Map<String, Object>> select(Query query) {
-        return getDataSourceReader().select(query);
+    public List<?> select(Query query) {
+        TraceFilterSplitter splitter = new TraceFilterSplitter(this.traceSpanSchema, this.traceTagIndexSchema);
+        splitter.split(query.getFilter());
+
+        return getTraceList(splitter.getExpression(),
+                            splitter.getIndexedTagFilters(),
+                            query.getInterval().getStartTime().toTimestamp(),
+                            query.getInterval().getEndTime().toTimestamp(),
+                            query.getOrderBy(),
+                            query.getLimit());
     }
 
     @Override
     public int count(Query query) {
-        return getDataSourceReader().count(query);
+        TraceFilterSplitter splitter = new TraceFilterSplitter(this.traceSpanSchema, this.traceTagIndexSchema);
+        splitter.split(query.getFilter());
+
+        return getTraceListSize(splitter.getExpression(),
+                                splitter.getIndexedTagFilters(),
+                                query.getInterval().getStartTime().toTimestamp(),
+                                query.getInterval().getEndTime().toTimestamp());
     }
 
     @Override
@@ -401,7 +423,7 @@ public class TraceJdbcReader implements ITraceReader {
 
                 if (((IdentifierExpression) left).getIdentifier().equals(kindFieldName)) {
                     if (right instanceof LiteralExpression) {
-                        String kindValue = (String) ((LiteralExpression) right).getValue();
+                        String kindValue = (String) ((LiteralExpression<?>) right).getValue();
                         isTrue = SpanKind.isRootSpan(kindValue);
                     }
                 }
@@ -427,12 +449,8 @@ public class TraceJdbcReader implements ITraceReader {
 
         @Override
         public boolean visit(LogicalExpression expression) {
-            if (expression instanceof LogicalExpression.NOT) {
-                // A little complicated to apply the optimization, ignore
-                return false;
-            }
-
-            return true;
+            // A little complicated to apply the optimization, ignore
+            return !(expression instanceof LogicalExpression.NOT);
         }
     }
 }
