+++ server/web-service/src/main/java/org/bithon/server/web/service/tracing/api/TraceApi.java
@@ -19,6 +19,8 @@ package org.bithon.server.web.service.tracing.api;
 import jakarta.validation.Valid;
 import org.bithon.component.commons.utils.Watch;
 import org.bithon.server.commons.time.TimeSpan;
+import org.bithon.server.storage.datasource.query.Limit;
+import org.bithon.server.storage.datasource.query.OrderBy;
 import org.bithon.server.storage.tracing.TraceSpan;
 import org.bithon.server.web.service.WebServiceModuleEnabler;
 import org.bithon.server.web.service.datasource.api.TimeSeriesQueryResult;
@@ -90,10 +92,8 @@ public class TraceApi {
             traceService.getTraceList(request.getExpression(),
                                       start,
                                       end,
-                                      request.getOrderBy(),
-                                      request.getOrder(),
-                                      request.getPageNumber(),
-                                      request.getPageSize())
+                                      new OrderBy(request.getOrderBy(), request.getOrder()),
+                                      new Limit(request.getPageSize(), request.getPageNumber() * request.getPageSize()))
         );
     }
